HSEARCH-3076 Move tests from ValueBridgeBaseIT to FieldBaseIT
Because that's where most ValueBridge-related tests currently are.
